minor: Fix variable size

+++ src/devices/wt0124.c
@@ -33,7 +33,7 @@ static int wt1024_callback(r_device *decoder, bitbuffer_t *bitbuffer)
 {
     data_t *data;
     uint8_t *b; // bits of a row
-    uint16_t sensor_rid;
+    uint8_t sensor_rid;
     int16_t value;
     float temp_c;
     uint8_t channel;
